reorganized null angle impl
